smartparking/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.7.2
boto3==1.35.5
botocore==1.35.5
CacheControl==0.14.0
cachetools==5.5.0
certifi==2023.7.22
cffi==1.17.0
charset-normalizer==3.3.0
cryptography==43.0.0
Django==4.2.6
django-cors-headers==4.2.0
djangorestframework==3.14.0
firebase-admin==6.5.0
google-api-core==2.19.1
google-api-python-client==2.141.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.17.2
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.5
grpcio-status==1.65.5
gunicorn==23.0.0
httplib2==0.22.0
idna==3.4
jmespath==1.0.1
msgpack==1.0.8
packaging==24.1
payos==0.1.4
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.9.0
pyparsing==3.1.2
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2023.3.post1
qrcode==7.4.2
requests==2.31.0
rsa==4.9
s3transfer==0.10.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.12.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
whitenoise==6.7.0
